pyproject.toml
[project]
name = "colibri"
version = "0.2.0"
description = "CoLibri (COntext LIBRary) - Local RAG system for technical books and notes"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Tobias Schelling" }
]
keywords = [
    "rag",
    "retrieval-augmented-generation",
    "knowledge-library",
    "llm",
    "knowledge-management",
    "semantic-search",
    "claude",
    "mcp",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Indexing",
]
requires-python = ">=3.11"
dependencies = [
    "lancedb>=0.17.0",
    "pymupdf4llm>=0.0.17",
    "mcp>=1.0.0",
    "python-frontmatter>=1.1.0",
    "watchdog>=4.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pandas>=2.0.0",
    # EPUB support
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.14.0",
    # Config
    "pyyaml>=6.0",
    # REST API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # TUI
    "textual>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "build>=1.2.1",
]

[project.scripts]
colibri = "colibri.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/colibri"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
